==> 1_content_collaborative_recommender.ipynb <==
# Begin Part I: Content Based Recommendation

import pandas as pd
import numpy as np

#Import data from the clean file 
df = pd.read_csv('./data/genome-scores.csv')
# example of record 
# movieId    tagId  Relevance
#  1           3    0.05775
df.head()

TableTag = df.copy()

tag_table = TableTag.pivot_table(values='relevance', index='movieId', columns='tagId')
# this generate a 10381 movie rows - 1128 tag column table

tag_table.to_csv(('./data/tag_table.csv')) # saving for the app

len(tag_table)

tag_table.head()

from sklearn.metrics.pairwise import cosine_similarity
cosine_sim = cosine_similarity(tag_table,tag_table) # create the matrix x1.x2/||x1||.||x2||

import pickle
pickle.dump(cosine_sim,open("./data/cosine_sim.p","wb")) ## saving for direct use to save time

cosine_sim.shape

cosine_sim[0,:] # looking at the first row

cosine_sim[0,1] # "distance" movie 1: toy story movie 2: jumanji

#Import the movie file
df2 = pd.read_csv('./data/movies.csv') # movieId - Title - Genre (adventure|animation....)

df2.head()

tag_table.tail()

dfmatch =  pd.DataFrame()
dfmatch['movieId'] = tag_table.index # this give index and movie number

dfmatch.tail()

df2.tail()

result=dfmatch.merge(df2,how='left',on='movieId')

result[(result['movieId'] == 131170)]

#Construct a reverse mapping of indices and movie titles
indices = pd.Series(result.index, index=result['title']) # match title and  0 - 10,380

indices.to_csv('./data/indices.csv')  
tag_table.to_csv('./data/movietagged.csv')
result.to_csv(('./data/moviemapped.csv')) 

df.head()

def content_tag_recommender(title, cosine_sim=cosine_sim, df=df, indices=indices):
    # Obtain the index of the movie that matches the title
    idx = indices[title]
    # Get the pairwsie similarity scores of all movies with that movie
    # And convert it into a list of tuples as described below
    sim_scores = list(enumerate(cosine_sim[idx]))
    #  with "Toy Story (1995)" in title
    # [(0, 0.9999999999999999), (1, 0.7743635326655509), (2, 0.7054215147240148), (3, 0.6530303898449995) ....

    # Sort the movies based on the cosine similarity scores
    sim_scores = sorted(sim_scores, key=lambda x: x[1], reverse=True)
    #[(0, 0.9999999999999999), (4331, 0.958830919528648), (2769, 0.9555216674015985), (2064, 0.9488923890883056)
     
    # Get the scores of the 100 most similar movies. Ignore the first movie. It is its self
    sim_scores = sim_scores[1:101] 
  
    # Get the movie indices
    movie_indices = [i[0] for i in sim_scores]
    #[4331, 2769, 2064, 5445, 9070, 7994, 4602, 3809 ......
    
    # Return the top 100 most similar movies
    return df['title'].iloc[movie_indices]

indices.head()

df3=content_tag_recommender("Wolverine, The (2013)",cosine_sim,result,indices)
# try the content function on wolverine movie

df3.to_csv('./data/contentreco.csv') # save the results

df3 # looks good

# End Part I Content Based Recommendation
# Begin Part II Collaborative Filtering

#Build the SVD based Collaborative filter. This creates all the prediction ratings
from surprise import SVD, Reader, Dataset

reader = Reader()
ratings = pd.read_csv('data/ratings.csv')
data = Dataset.load_from_df(ratings[['userId', 'movieId', 'rating']], reader)
data.split(n_folds=5)
svd = SVD()
trainset = data.build_full_trainset()
svd.fit(trainset)

import pickle
pickle.dump(svd,open("./data/svdpred.p","wb"))#saving the svd values because it consume lot of time,so we can refer this file for svd values later

svd.predict(1, 29).est # what would be the rating for viewer 1 on movie 29

df4 = pd.read_csv('./data/contentreco.csv',names = ['TagmovieId','title'],header = None)

df3.head() # the 100 movies

df100 = pd.DataFrame(df3)
df100['TagmovieId'] = df3.index 
df100.head(20) # introducing tagmovieid which is a continuous sequence of all rated films

df4.head() # when loading the file we can get the 100 movie under the format below

result.tail() #correspondence TagmovieId continuous sequence for the movie <->the movieId

df5 = pd.merge(df4, result, left_index=True, on=['title'])
# 	TagmovieId 	title 	movieId 	genres
#8700 	8700 	X-Men Origins: Wolverine (2009) 	68319 	Action|Sci-Fi|Thriller

df5.head() # getting a table showing the correspondence for the 100 movies

def cross_rating(UID,MID):
    rating = ratings[(ratings['userId'] == UID) & (ratings['movieId'] == MID)]['rating']
    if rating.size == 0: 
        rate = 0
    elif rating.size == 1: 
        rate = rating.values[0]   
    return rate

cross_rating(248,77561) # getting the rating viewer 248 gave to movie 77561

cross_rating(248,68319) # getting the rating viewer 248 gave to movie 68319 -> no rating

USER = 248
df5['givenrating'] = df5.apply(lambda x: cross_rating(USER,x['movieId']), axis=1)
# getting all the rating viewer 248 gave to the 100 movies

df5.head(1)

df5['predrating'] = df5.apply(lambda x: svd.predict(USER, x['movieId']).est, axis=1)
# predict all the rating viewer 248 would give to the 100 movies

df5.head()

df5[(df5['givenrating'] != 0)] # compare actual and predicted ratings

df5['rating'] = df5.apply(lambda x: x['givenrating'] if x['givenrating'] != 0 else(x['predrating']) , axis=1)
# rating fields contains the predicted rating or the rating given if it exists 

df5[(df5['givenrating'] != 0)]

df6 = df5.sort_values(['rating'],ascending=[0])
# sort in the highest order

cols = [1,2,4,6]
df7 = df6[df6.columns[cols]].head(15) # just show title, moviedId,givenrating and rating
df7.head()

# we will exclude the movies with givenrating <> 0 for recommendations
df7[(df7['givenrating'] != df7['rating'])].head(5)

# Here are 5 movie recommendation for viewer 248 similar to movie "Wolverine, The (2013)"

df5.head()

result.head()
