==> 3_hybrid_recommender.ipynb <==
from movie_recommender import *

print("enter your movie")
print("""exemples:
        Wolverine, The (2013)
        Magnificent Seven, The (1960)
        True Grit (2010)
        Lord of the Rings: The Two Towers, The (2002)
        Dark Knight, The (2008) 
            """)
movie = input()

content_tag_recommender(movie)

print("enter your viewer ID: (exemples 1 4 248 300)")
viewer = input()

viewer = int(viewer)
collaborative_recommender(viewer,movie)

print("enter your viewer ID exemples 1 4 248 300")
viewer = input()

viewer = int(viewer)
collaborative_recommender(viewer,movie)

print("enter your viewer ID: (exemples 1 4 248 300)")
viewer = input()

viewer = int(viewer)
collaborative_recommender(viewer,movie)

#end
